requirements.txt
django-cms
django-crispy-forms
raven
reportlab
South
django-sendfile
python-memcached
django-tinymce
